Example of input data using local data

Input example from disk data. See details

Format

an array with two layers (temp_dat) a matrix (rain_dat), a data.frame (innlandet), a sf object (lonlatsf), a list ("commonbean")

Source

Funk, C. et al. (2015). The climate hazards infrared precipitation with stations—a new environmental record for monitoring extremes. Scientific Data, 2, 150066. doi:10.1038/sdata.2015.66

van Etten, J. et. al. (2019). PNAS 116(10) 4194-4199 doi:10.1073/pnas.1813720116

Wan Z, Hook S, Hulley G (2015) MYD11A1 MODIS/Aqua Land Surface Temperature/Emissivity 8-Day L3 Global 1km SIN Grid V006 doi:10.5067/MODIS/MYD11A2.006 .

Details

commonbean: a list with three data sets 1) on-farm variety performance data, 2) chirps data, and 3) modis data.

rain_dat: matrix with precipitation from CHIRPS. An excerpt to represent an example of the input data in rainfall() or

get_timeseries() when a matrix is provided.

rain_dat: matrix with precipitation from CHIRPS. An excerpt to represent an example of the input data in rainfall() or get_timeseries() when a matrix is provided.

innlandet: a data.frame with maximum and minimum temperature for a random point in the Innlandet county in Norway, spanning from "2019-01-01" to "2019-07-01"

lonlatsf: a 'sf' object with 'POINT' geometry with five random points around the municipality of Sinop, Brazil

temp_dat and rain_dat, rows represents the coordinates for the given lonlat provided in raster::extract() and colunms represents the dates from the observed precipitation/temperature.