Time series climate data

General functions and methods to concatenate climate data across a time series

get_timeseries(object, day.one, ...)

# S3 method for default
get_timeseries(
  object,
  day.one,
  span = NULL,
  last.day = NULL,
  data.from = "nasapower",
  ...
)

# S3 method for sf
get_timeseries(
  object,
  day.one,
  span = NULL,
  last.day = NULL,
  data.from = "nasapower",
  ...
)

# S3 method for matrix
get_timeseries(object, day.one, span = NULL, last.day = NULL, ...)

# S3 method for array
get_timeseries(object, day.one, span = NULL, last.day = NULL, ...)

Arguments

object
a data.frame (or any other object that can be coerced to data.frame) with geographical coordinates (lonlat), or an object of class sf with geometry 'POINT' or 'POLYGON', or a named matrix with climate data, or an array with two dimensions for max and min temperature. See details.

day.one
a vector of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the starting day to capture the climate data

...
additional arguments passed to methods. See details.

span
an integer or a vector with integers (optional if last.day is given) for the length of the time series to be captured

last.day
optional to span, an object of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the last day of the time series

data.from
character, for the source of climate data. See details.

Value

A list with class clima_ls with data.frame(s) with the class clima_df

Details

The default method and the sf method assumes that the climate data will be fetched from an remote (cloud) data.from.

The matrix method assumes that the climate data was previously handled and will be inputted in the format of a named matrix. See help("modis", "climatrends") for examples.

Available remote sources to pass data.from: "nasapower"

Additional arguments:

pars: character vector of solar, meteorological or climatology parameters to download. See help("parameters", "nasapower") when data.from = "nasapower".

days.before: an integer for the number of days before day.one to be included in the timespan.

Examples

# Using local sources
# an array with temperature data
data("temp_dat", package = "climatrends")

set.seed(9271)
span <- as.integer(runif(10, 6, 15))

get_timeseries(temp_dat, "2013-10-28", span = span)
#> [[1]]
#>        id       date value
#>     <int>     <date> <dbl>
#> 1:      1 2013-10-28 33.20
#> 2:      1 2013-10-29 33.40
#> 3:      1 2013-10-30 33.50
#> 4:      1 2013-10-31 33.50
#> 5:      1 2013-11-01 33.50
#> ---                       
#> 87:    10 2013-11-01 28.20
#> 88:    10 2013-11-02 28.20
#> 89:    10 2013-11-03 28.10
#> 90:    10 2013-11-04 28.10
#> 91:    10 2013-11-05 28.10
#> 
#> [[2]]
#>        id       date value
#>     <int>     <date> <dbl>
#> 1:      1 2013-10-28  8.20
#> 2:      1 2013-10-29  8.20
#> 3:      1 2013-10-30  8.30
#> 4:      1 2013-10-31  8.30
#> 5:      1 2013-11-01  8.30
#> ---                       
#> 87:    10 2013-11-01  3.20
#> 88:    10 2013-11-02  3.40
#> 89:    10 2013-11-03  3.50
#> 90:    10 2013-11-04  3.70
#> 91:    10 2013-11-05  3.90
#> 
#> attr(,"class")
#> [1] "clima_ls" "list"    

# matrix with precipitation data
data("rain_dat", package = "climatrends")

get_timeseries(rain_dat, "2013-10-28", span = span)
#> [[1]]
#>        id       date value
#>     <int>     <date> <dbl>
#> 1:      1 2013-10-28  0.00
#> 2:      1 2013-10-29  0.00
#> 3:      1 2013-10-30  0.00
#> 4:      1 2013-10-31  0.00
#> 5:      1 2013-11-01  0.00
#> ---                       
#> 87:    10 2013-11-01  0.00
#> 88:    10 2013-11-02  0.00
#> 89:    10 2013-11-03  0.00
#> 90:    10 2013-11-04  0.00
#> 91:    10 2013-11-05  0.00
#> 
#> attr(,"class")
#> [1] "clima_ls" "list"    

########################################################
# \donttest{
library("nasapower")
library("sf")
# Fetch data from NASA POWER using 'sf' method
data("lonlatsf", package = "climatrends")

g <- get_timeseries(lonlatsf, 
                    day.one = "2018-05-16", 
                    last.day = "2018-05-30",
                    pars = c("PRECTOT", "T2M", "T10M"))
#> Getting climate data from NASA POWER 
#> Error: PRECTOT is/are not valid in 'pars'.
#> Check that the 'pars', 'community' and 'temporal_api' align.

g
#> Error in eval(expr, envir, enclos): object 'g' not found
# }