Rainfall indices

Compute rainfall indices over a timespan

rainfall(
  object,
  day.one = NULL,
  span = 150,
  timeseries = FALSE,
  intervals = 5,
  ...
)

Arguments

object
a numeric vector of geographic coordinates (lonlat) or a matrix containing the precipitation data. When lonlat is used, the function makes a call to nasapower::get_power to fetch and concatenate environmental data from NASA POWER (https://power.larc.nasa.gov/)

day.one
a vector of class Date for the starting date to capture the environmental data (YYYY-MM-DD)

span
an integer or a vector with integers for the duration of the timespan to be captured

timeseries
logical, FALSE for a single point time series observation or TRUE for a time series based on intervals

intervals
integer no lower than 5, for the days intervals when timeseries = TRUE

...
additional arguments passed to get_power methods

Value

A dataframe with rainfall indices:

MLDS

maximum length of consecutive dry day, rain < 1 mm (days)

MLWS

maximum length of consecutive wet days, rain >= 1 mm (days)

R10mm

number of heavy precipitation days 10 >= rain < 20 mm (days)

R20mm

number of very heavy precipitation days rain >= 20 (days)

Rx1day

maximum 1-day precipitation (mm)

Rx5day

maximum 5-day precipitation (mm)

R95p

total precipitation when rain > 95th percentile (mm)

R99p

total precipitation when rain > 99th percentile (mm)

Rtotal

total precipitation (mm) in wet days, rain >= 1 (mm)

SDII

simple daily intensity index, total precipitation divided by the number of wet days (mm/days)

When timeseries = TRUE, an id is created, which is the index for the rownames of the provided object.
References

Aguilar E., et al. (2005). Journal of Geophysical Research, 110(D23), D23107.
https://doi.org/10.1029/2005JD006119

Kehel Z., et al. (2016). Identifying Climate Patterns during the Crop-Growing Cycle from 30 Years of CIMMYT Elite Spring Wheat International Yield Trials. In: Applied Mathematics and Omics to Assess Crop Genetic Resources for Climate Change Adaptive Traits (eds Bari A., Damania A. B., Mackay M., Dayanandan S.), pp. 151–174. CRC Press.

Sparks A. H. (2018). Journal of Open Source Software, 3(30), 1035.
https://doi.org/10.21105/joss.01035

See also

Examples

# \donttest{
library("nasapower")

# random points within bbox(11, 12, 55, 58)
set.seed(123)
lonlat <- data.frame(lon = runif(3, 11, 12),
                     lat = runif(3, 55, 58))

# random dates within 2018-05-15 and 2018-05-20
set.seed(321)
dates <- as.integer(runif(3, 17666, 17670))
dates <- as.Date(dates, origin = "1970-01-01")

# calculate rainfall for the first 50 days after day.one
rainfall(lonlat,
         day.one = dates,
         span = 50)

#> # A tibble: 3 x 10
#>    MLDS  MLWS R10mm R20mm Rx1day Rx5day  R95p  R99p Rtotal  SDII
#>   <int> <int> <int> <int>  <dbl>  <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1    15     6     1     0  11.0   26.9   27.1 11.0    61.4  5.58
#> 2    15     6     1     0  11.1   27     27.2 11.1    64.3  5.85
#> 3    12     2     0     0   4.88   9.31  12.3  4.88   25.2  4.19

# include the first 15 days before day.one (residual precipitation)
rainfall(lonlat,
         day.one = dates,
         span = 50,
         days.before = 15)

#> # A tibble: 3 x 10
#>    MLDS  MLWS R10mm R20mm Rx1day Rx5day  R95p  R99p Rtotal  SDII
#>   <int> <int> <int> <int>  <dbl>  <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1    24     6     1     0  11.0   26.9   33.0 11.0    64.8  5.40
#> 2    24     6     1     0  11.1   27     33.6 11.1    67.9  5.66
#> 3    15     2     0     0   4.88   9.31  16.4  4.88   36.0  3.60

# rainfall indices over a time series          
rainfall(lonlat,
         day.one = dates,
         span = 50,
         timeseries = TRUE,
         intervals = 7)

#> # A tibble: 210 x 4
#>       id date       index  value
#>    <int> <date>     <chr>  <dbl>
#>  1     1 2018-05-18 MLDS    7   
#>  2     1 2018-05-18 MLWS    0   
#>  3     1 2018-05-18 R10mm   0   
#>  4     1 2018-05-18 R20mm   0   
#>  5     1 2018-05-18 Rx1day  0.16
#>  6     1 2018-05-18 Rx5day  0.17
#>  7     1 2018-05-18 R95p    0.16
#>  8     1 2018-05-18 R99p    0.16
#>  9     1 2018-05-18 Rtotal  0.2 
#> 10     1 2018-05-18 SDII    0   
#> # … with 200 more rows

# }