Temperature indices

Methods to compute temperature indices over a time series

temperature(object, ...)

# S3 method for default
temperature(
  object,
  day.one,
  span = NULL,
  timeseries = FALSE,
  intervals = 5,
  ...
)

# S3 method for array
temperature(
  object,
  day.one,
  span = NULL,
  timeseries = FALSE,
  intervals = 5,
  ...
)

# S3 method for sf
temperature(
  object,
  day.one,
  span = NULL,
  timeseries = FALSE,
  intervals = 5,
  as.sf = TRUE,
  ...
)

# S3 method for clima_ls
temperature(object, timeseries = FALSE, intervals = 5, ...)

Arguments

object
a data.frame (or any object that can be coerced to data.frame) with geographical coordinates (lonlat), or an object of class sf with geometry 'POINT' or 'POLYGON', or an array with two dimensions containing the temperature data or a clima_ls with day and night temperature, in that order. See details

...
additional arguments passed to methods. See details

day.one
a vector of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the starting day to capture the climate data

span
an integer or a vector with integers (optional if last.day is given) for the length of the time series to be captured

timeseries
logical, FALSE for a single point time series observation or TRUE for a time series based on intervals

intervals
integer (no lower than 5), for the days intervals when timeseries = TRUE

as.sf
logical, to return an object of class 'sf'

Value

A dataframe with temperature indices:

maxDT

maximun day temperature (degree Celsius)

minDT

minimum day temperature (degree Celsius)

maxNT

maximun night temperature (degree Celsius)

minNT

minimum night temperature (degree Celsius)

DTR

diurnal temperature range (mean difference between DT and NT (degree Celsius))

SU

summer days, number of days with maximum temperature > 30 (degree Celsius)

TR

tropical nights, number of nights with maximum temperature > 25 (degree Celsius)

CFD

consecutive frosty days, number of days with temperature bellow 0 degree Celsius

WSDI

maximum warm spell duration, consecutive days with temperature > 90th percentile

CSDI

maximum cold spell duration, consecutive nights with temperature < 10th percentile

T10p

the 10th percentile of night tempeture (degree Celsius)

T90p

the 90th percentile of day tempeture (degree Celsius)

Details

The array method assumes that object contains climate data provided from a local source; this requires an array with two dimensions, 1st dimension contains the day temperature and 2nd dimension the night temperature, see help("modis", package = "climatrends") for an example on input structure.

The default method and the sf method assumes that the climate data will e fetched from a remote (cloud) source that be adjusted using the argument data.from.

When timeseries = TRUE, an id is created, which is the index for the rownames of the inputted object.

Additional arguments:

last.day: optional to span, an object of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the last day of the time series

data.from: character for the source of remote data. Current remote source is: 'nasapower'

pars: character vector for the temperature data to be fetched. If data.from is 'nasapower', the temperature can be adjusted to 2 m, the default, c("T2M_MAX", "T2M_MIN") or 10 m c("T10M_MAX", "T10M_MIN")

days.before: optional, an integer for the number of days before day.one to be included in the timespan.

References

Aguilar E., et al. (2005). Journal of Geophysical Research, 110(D23), D23107.
https://doi.org/10.1029/2005JD006119

See also

Examples

# Using local sources
data("modis", package = "climatrends")

temperature(modis,
            day.one = "2013-10-28",
            span = 12)

#>     maxDT minDT maxNT minNT   DTR    SU    TR   CFD  WSDI  CSDI  T10p  T90p
#>     <dbl> <dbl> <dbl> <dbl> <int> <int> <int> <int> <int> <int> <dbl> <dbl>
#> 1:  33.50 33.20  8.40  8.20    25    12     0     0     5     2  8.21 33.50
#> 2:  34.90 34.60  9.50  8.90    25    12     0     0     6     2  8.91 34.90
#> 3:  32.30 31.70  8.50  8.00    23    12     0     0     8     2  8.01 32.00
#> 4:  27.50 26.90  2.10  1.50    25     0     0     0     2     5  1.60 27.40
#> 5:  32.80 32.00  6.40  5.70    26    12     0     0     2     3  5.70 32.58
#> 6:  27.60 26.90  6.50  5.90    20     0     0     0     2     2  6.01 27.40
#> 7:  32.80 32.00  6.40  5.70    26    12     0     0     2     3  5.70 32.58
#> 8:  31.90 31.00  4.90  4.60    26    12     0     0     2     2  4.61 31.57
#> 9:  32.60 31.30  5.10  4.30    27    12     0     0     2     3  4.30 32.27
#> 10: 28.70 28.00  4.40  2.70    24     0     0     0     2     2  2.81 28.50

# \donttest{
#####################################################

# Using remote sources of climate data

data("lonlatsf", package = "climatrends")

# some random dates provided as integers and coerced to Dates internally
set.seed(2718279)
dates <- as.integer(runif(5, 17660, 17675))

# get temperature indices for 30 days after day.one
# return a data.frame
temp1 <- temperature(lonlatsf,
                     day.one = dates,
                     span = 30,
                     as.sf = FALSE)

#> Getting climate data from NASA POWER 

temp1

#>    maxDT minDT maxNT minNT   DTR    SU    TR   CFD  WSDI  CSDI  T10p  T90p
#>    <dbl> <dbl> <dbl> <dbl> <int> <int> <int> <int> <int> <int> <dbl> <dbl>
#> 1: 17.25  3.35 10.62 -4.01     7     0     0     4     2     2 -2.42 15.70
#> 2: 21.68  9.54 12.97  0.41    10     0     0     0     1     2  1.15 20.81
#> 3: 27.51 14.61 16.31  4.32    11     0     0     0     1     2  5.91 26.39
#> 4: 22.59 11.12 13.06  0.47    10     0     0     0     3     2  2.28 21.98
#> 5: 24.81 11.74 14.68  0.10    11     0     0     0     2     2  2.63 23.74

# get temperature indices from "2010-12-01" to "2011-01-31"
temp2 <- temperature(lonlatsf,
                     day.one = "2010-12-01",
                     last.day = "2011-01-31",
                     as.sf = FALSE)

#> Getting climate data from NASA POWER 

temp2

#>    maxDT  minDT maxNT  minNT   DTR    SU    TR   CFD  WSDI  CSDI   T10p  T90p
#>    <dbl>  <dbl> <dbl>  <dbl> <int> <int> <int> <int> <int> <int>  <dbl> <dbl>
#> 1: -1.31 -20.80 -6.11 -24.51     4     0     0    62     2     4 -21.95 -3.06
#> 2: -0.82 -19.68 -5.69 -25.47     6     0     0    62     2     4 -23.02 -2.64
#> 3: -0.10 -17.83 -1.74 -27.12     6     0     0    62     3     2 -20.00 -0.88
#> 4: -0.06 -18.38 -5.35 -24.35     6     0     0    62     2     4 -22.39 -1.80
#> 5: -0.18 -18.92 -3.72 -27.22     6     0     0    62     3     2 -20.65 -3.27

# indices from "2010-12-01" to "2011-01-31" with intervals of 7 days
temp3 <- temperature(lonlatsf,
                     day.one = "2010-12-01",
                     last.day = "2011-01-31",
                     timeseries = TRUE,
                     intervals = 7,
                     as.sf = FALSE)

#> Getting climate data from NASA POWER 

temp3

#>         id   lon   lat       date index  value
#>      <int> <dbl> <dbl>     <date> <chr>  <dbl>
#> 1:       1  8.35 61.86 2010-12-01 maxDT -13.51
#> 2:       1  8.35 61.86 2010-12-01 minDT -18.89
#> 3:       1  8.35 61.86 2010-12-01 maxNT -16.03
#> 4:       1  8.35 61.86 2010-12-01 minNT -22.16
#> 5:       1  8.35 61.86 2010-12-01   DTR   3.41
#> ---                                           
#> 476:     5 11.21 61.32 2011-01-19   CFD   7.00
#> 477:     5 11.21 61.32 2011-01-19  WSDI   1.00
#> 478:     5 11.21 61.32 2011-01-19  CSDI   1.00
#> 479:     5 11.21 61.32 2011-01-19  T10p -13.34
#> 480:     5 11.21 61.32 2011-01-19  T90p  -1.52

# }