Growing degree-days

Compute number of days required to reach Growing degree-days. GDD is a heuristic tool in phenology that measures heat accumulation and is used to predict plant and animal development rates. Growing degree-days are calculated by taking the integral of warmth above a base temperature.

GDD(object, day.one, degree.days = NULL, base = 10, ...)

# S3 method for default
GDD(
  object,
  day.one,
  degree.days = NULL,
  base = 10,
  span = NULL,
  return = "ndays",
  ...
)

# S3 method for array
GDD(object, day.one, degree.days = NULL, base = 10, return = "ndays", ...)

# S3 method for sf
GDD(
  object,
  day.one,
  degree.days = NULL,
  base = 10,
  span = NULL,
  return = "ndays",
  as.sf = TRUE,
  ...
)

Arguments

object
a data.frame (or object that can be coerced to data.frame) with geographical coordinates (lonlat), or an object of class sf with geometry 'POINT' or 'POLYGON', or an array with two dimensions containing the temperature data, see details.

day.one
a vector of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the starting day to capture the climate data

degree.days
an integer for the degree-days required by the organism (look for the physiology of the focal organism). Optional if return = "daily"

base
an integer for the base temperature

...
additional arguments passed to methods. See details.

span
an integer or a vector with integers (optional if last.day is given) for the length of the time series to be captured

return
character (one of "ndays" or "daily") to select if the function returns the number of days to reach the GDD or the raw daily values of the GDD

as.sf
logical, to return an object of class 'sf'

Value

The number of days which were required to reach the growing degree-days, or the degree days in each day during the time series

Details

The array method assumes that object contains climate data provided from a local source; this requires an array with two dimensions, 1st dimension contains the day temperature and 2nd dimension the night temperature, see help("modis", "climatrends") for an example on input structure.

The default method and the sf method assumes that the climate data will be fetched from an remote (cloud) source.

Additional arguments:

last.day: optional to span, an object of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the last day of the time series

source: character for the source of remote data. Current remote source is: 'nasapower'

pars: character vector for the temperature data to be fetched. If source is 'nasapower'. The temperature can be adjusted to 2 m, the default, c("T2M_MAX", "T2M_MIN") or 10 m c("T10M_MAX", "T10M_MIN")

References

Prentice I. C., et al. (1992) Journal of Biogeography, 19(2), 117.
https://doi.org/10.2307/2845499

See also

Examples

# Using local sources
data("modis", package = "climatrends")

GDD(modis,
    day.one = "2013-10-28",
    degree.days = 100,
    base = 5)

######################################
# \donttest{
# Using remote sources
set.seed(123)
# random geographic locations around bbox(11, 12, 55, 58)
lonlat <- data.frame(lon = runif(3, 11, 12),
                     lat = runif(3, 55, 58))

set.seed(321)
# random dates around 2018-05-15 and 2018-05-20
dates <- as.integer(runif(3, 17666, 17670))

# Calculate the days required for the plants in these plots to reach the
# maturity. Here the plant species requires ~1300 degree days for it.
GDD(lonlat,
    day.one = dates,
    degree.days = 1300,
    span = 150,
    base = 5)

######################################

# Objects of class 'sf'
data("lonlatsf", package = "climatrends")

dates <- as.Date(16150, origin = "1970-01-01")

GDD(lonlatsf,
    day.one = "2014-03-21",
    last.day = "2014-06-30",
    degree.days = 200,
    base = 5)

# }