Growing degree-days

This a heuristic tool in phenology that measures heat accumulation and is used to predict plant and animal development rates. Growing degree-days are calculated by taking the integral of warmth above a base temperature.

GDD(..., base = 10)

# S3 method for default
GDD(tmax, tmin, dates = NULL, ..., base = 10)

# S3 method for data.frame
GDD(object, day.one, ..., base = 10)

# S3 method for array
GDD(object, day.one, ..., base = 10)

# S3 method for sf
GDD(object, day.one, ..., base = 10, as.sf = TRUE)

# S3 method for clima_ls
GDD(object, ..., base = 10)

Arguments

...
additional arguments passed to methods. See details

base
an integer for the minimum temperature for growth

tmax
a numeric vector with the maximum temperature

tmin
a numeric vector with the minimum temperature

dates
optional, a character (or Date or numeric) vector for the dates in tmax and tmin

object
a data.frame with geographical coordinates (lonlat), or an object of class sf with geometry 'POINT' or 'POLYGON', or an array with two dimensions containing the temperature data or a clima_ls with maximum and minimum temperature, in that order. See details

day.one
a vector of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the starting day to capture the climate data

as.sf
logical, to return an object of class 'sf'

Value

The number of days to reach the accumulated degree.days or the daily degree-days as defined with the argument return.as

Details

Additional arguments:

return.as character (one of, the default, "gdd" or "ndays") to select if the function returns the number of days to reach the accumulated degree.days or the daily values of the GDD

degree.days an integer for the accumulated degree-days required by the organism (as per the physiology of the focal organism). Optional if return.as = "gdd"

equation character to specify the equation to be used, one of "default", "variant_a" or "variant_b". See Equations below

last.day: an object (optional to span) of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the last day of the time series

span: an integer (optional to last.day) or a vector with integers (optional if last.day is given) for the length of the time series to be captured

pars: character vector for the temperature data to be fetched. If data.from is 'nasapower'. The temperature can be adjusted to 2 m, the default, c("T2M_MAX", "T2M_MIN") or 10 m c("T10M_MAX", "T10M_MIN")

S3 Methods:

The array method assumes that object contains climate data available in your R section; this requires an array with two dimensions, 1st dimension contains the day temperature and 2nd dimension the night temperature, see help("temp_dat", package = "climatrends") for an example on input structure.

The data.frame method and the sf method assumes that the climate data will e fetched from a remote (cloud) source that be adjusted using the argument data.from

Equations:

The "default" equation uses the average of the daily maximum (Tmax) and minimum (Tmin) temperatures compared to a base temperature (Tbase). If Tmin is below Tbase there are two variants:

"variant_a": set Tmean = Tbase if (Tmax + Tmin) / 2 < Tbase

"variant_b": set Tmin = Tbase if Tmin < Tbase

References

Prentice I. C., et al. (1992) Journal of Biogeography, 19(2), 117.
https://doi.org/10.2307/2845499

See also

Examples

data("innlandet", package = "climatrends")

tmax <- innlandet[, "tmax"]
tmin <- innlandet[, "tmin"]

# using the default equation, returns negative values
GDD(tmax, tmin, base = 2)

# set the equation variant_a
# Tmean = Tbase if (Tmax + Tmin) / 2 < Tbase
GDD(tmax, tmin, base = 2, equation = "variant_a")

#####################################################

# return as the number of days required to reach a certain 
# accumulated GDD
data("temp_dat", package = "climatrends")

GDD(temp_dat, day.one = "2013-10-27", degree.days = 90, return.as = "ndays")