Time series climate data

General functions and methods to concatenate climate data across a time series.

get_timeseries(object, day.one, span = NULL, last.day = NULL, ...)

# S3 method for default
get_timeseries(
  object,
  day.one,
  span = NULL,
  last.day = NULL,
  source = "nasapower",
  ...
)

# S3 method for sf
get_timeseries(
  object,
  day.one,
  span = NULL,
  last.day = NULL,
  source = "nasapower",
  ...
)

# S3 method for matrix
get_timeseries(object, day.one, span = NULL, last.day = NULL, ...)

Arguments

object
a data.frame (or any othwer object that can be coerced to data.frame) with geographical coordinates (lonlat), or an object of class sf with geometry 'POINT' or 'POLYGON', or a named matrix with climate data. See details.

day.one
a vector of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the starting day to capture the climate data

span
an integer or a vector with integers (optional if last.day is given) for the length of the time series to be captured

last.day
optional to span, an object of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the last day of the time series

...
additional arguments passed to methods. See details.

source
character, for the source of climate data. See details.

Value

An object with time series climate data for the chosen period

Details

The default method and the sf method assumes that the climate data will be fetched from an remote (cloud) source.

The matrix method assumes that the climate data was previously handled and will be inputted in the format of a named matrix. See help("modis", "climatrends") for examples.

Available remote sources are: "nasapower"

Additional arguments:

pars: character vector of solar, meteorological or climatology parameters to download. See help("parameters", "nasapower") when source = "nasapower".

days.before: an integer for the number of days before day.one to be included in the timespan.

Examples

# Using local sources
data("modis", package = "climatrends")

day <- as.Date("2013-10-28", format = "%Y-%m-%d")

get_timeseries(chirp, day, span = c(10,11))

########################################################
# \donttest{
# Fetch data from NASA POWER using 'sf' method

data("lonlatsf", package = "climatrends")

do <- as.Date(17667, origin = "1970-01-01")

g <- get_timeseries(lonlatsf,
                    day.one = do,
                    span = 10,
                    pars = c("PRECTOT", "T2M", "T10M"))

g
# }