Rainfall indices

Methods to compute rainfall indices over a time series

rainfall(object, ...)

# S3 method for default
rainfall(object, dates = NULL, ..., timeseries = FALSE, intervals = 5)

# S3 method for data.frame
rainfall(object, day.one, span = NULL, ..., timeseries = FALSE, intervals = 5)

# S3 method for matrix
rainfall(object, day.one, span = NULL, ..., timeseries = FALSE, intervals = 5)

# S3 method for sf
rainfall(
  object,
  day.one,
  span = NULL,
  ...,
  timeseries = FALSE,
  intervals = 5,
  as.sf = TRUE
)

Arguments

object
a numeric vector with precipitation data or a data.frame with geographical coordinates (lonlat), or an object of class sf with geometry 'POINT' or 'POLYGON', or a named matrix with precipitation data. See details.

...
additional arguments passed to methods. See details.

dates
optional, a character (or Date or numeric) vector for the dates in object in the dafault method

timeseries
logical, FALSE for a single point time series observation or TRUE for a time series based on intervals

intervals
integer (no lower than 5), for the days intervals when timeseries = TRUE

day.one
a vector of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the starting day to capture the climate data

span
an integer or a vector with integers (optional if last.day is given) for the length of the time series to be captured

as.sf
logical, to return an object of class 'sf'

Value

A dataframe with rainfall indices:

MLDS

maximum length of consecutive dry day, rain < 1 mm (days)

MLWS

maximum length of consecutive wet days, rain >= 1 mm (days)

R10mm

number of heavy precipitation days 10 >= rain < 20 mm (days)

R20mm

number of very heavy precipitation days rain >= 20 (days)

Rx1day

maximum 1-day precipitation (mm)

Rx5day

maximum 5-day precipitation (mm)

R95p

total precipitation when rain > 95th percentile (mm)

R99p

total precipitation when rain > 99th percentile (mm)

Rtotal

total precipitation (mm) in wet days, rain >= 1 (mm)

SDII

simple daily intensity index, total precipitation divided by the number of wet days (mm/days)

Details

The matrix method assumes that object contains climate data available in your R section; see help("rain_dat", package = "climatrends") for an example on input structure.

The default method and the sf method assumes that the climate data will e fetched from a remote (cloud) source that be adjusted using the argument data.from.

When timeseries = TRUE, an id is created, which is the index for the rownames of the inputted object.

Additional arguments:

last.day: optional to span, an object of class Date or any other object that can be coerced to Date (e.g. integer, character YYYY-MM-DD) for the last day of the time series

data.from: character for the source of remote data. Current remote source is: 'nasapower'

pars: character vector for the precipitation data to be fetched. If data.from is 'nasapower', the default precipitation parameter is "PRECTOT".

days.before: optional, an integer for the number of days before day.one to be included in the timespan.

When timeseries = TRUE, an id is created, which is the index for the rownames of the inputted object.

References

Aguilar E., et al. (2005). Journal of Geophysical Research, 110(D23), D23107.
https://doi.org/10.1029/2005JD006119

Examples

# A vector with precipitation data
set.seed(987219)
rain <- runif(50, min = 0, max = 6)

rainfall(rain)

# Return as timeseries with intervals of 7 days
dates <- 17650:17699
rainfall(rain, dates = dates, timeseries = TRUE, intervals = 7)

######################################################

# the matrix method
data("rain_dat", package = "climatrends")

rainfall(rain_dat,
         day.one = "2013-10-28",
         span = 12)

#####################################################
# \donttest{
# Using remote sources of climate data
library("nasapower")
data("lonlatsf", package = "climatrends")

# some random dates provided as integers and coerced to Dates internally
set.seed(2718279)
dates <- as.integer(runif(5, 17660, 17675))

# get precipitation indices for 30 days after day.one
# return a data.frame
rain1 <- rainfall(lonlatsf,
                  day.one = dates,
                  span = 30,
                  as.sf = FALSE)
rain1

# get precipitation indices from "2010-12-01" to "2011-01-31"
rain2 <- rainfall(lonlatsf,
                  day.one = "2010-12-01",
                  last.day = "2011-01-31",
                  as.sf = FALSE)
rain2
# }